Move spectrum cli to workflows
