+++ asapdiscovery-spectrum/asapdiscovery/spectrum/tests/test_struct_align.py
@@ -7,8 +7,6 @@ from asapdiscovery.spectrum.align_seq_match import (
     save_pymol_seq_align,
 )
 from asapdiscovery.spectrum.calculate_rmsd import rmsd_alignment, save_alignment_pymol
-from asapdiscovery.spectrum.cli import spectrum as cli
-from click.testing import CliRunner
 
 
 def click_success(result):
@@ -75,161 +73,3 @@ def test_pymol_seq_align(protein_path, tmp_path):
         session_save=pse_out,
     )
     assert pse_out.exists()
-
-
-@pytest.mark.skipif(os.getenv("SKIP_EXPENSIVE_TESTS"), reason="Expensive tests skipped")
-def test_struct_alignment_single_pdb(blast_csv_path, protein_path, tmp_path):
-    runner = CliRunner()
-    result = runner.invoke(
-        cli,
-        [
-            "struct-alignment",
-            "-f",
-            blast_csv_path,
-            "--pdb-file",
-            protein_path,
-            "--pdb-align",
-            protein_path,
-            "--pymol-save",
-            tmp_path / "file.pse",
-            "--chain",
-            "both",
-            "--color-by-rmsd",
-        ],
-    )
-    assert click_success(result)
-
-
-@pytest.mark.skipif(os.getenv("SKIP_EXPENSIVE_TESTS"), reason="Expensive tests skipped")
-def test_struct_alignment_one_chain(
-    blast_csv_path, protein_path, protein_apo_path, tmp_path
-):
-    runner = CliRunner()
-    result = runner.invoke(
-        cli,
-        [
-            "struct-alignment",
-            "-f",
-            blast_csv_path,
-            "--pdb-file",
-            protein_path,
-            "--pdb-align",
-            protein_apo_path,
-            "--output-dir",
-            tmp_path,
-            "--pymol-save",
-            tmp_path / "file.pse",
-            "--chain",
-            "A",
-            "--color-by-rmsd",
-        ],
-    )
-    assert click_success(result)
-
-
-@pytest.mark.skipif(os.getenv("SKIP_EXPENSIVE_TESTS"), reason="Expensive tests skipped")
-def test_struct_alignment_struct_dir(
-    blast_csv_path, protein_path, structure_dir, tmp_path
-):
-    runner = CliRunner()
-    struct_dir, _ = structure_dir
-    result = runner.invoke(
-        cli,
-        [
-            "struct-alignment",
-            "-f",
-            blast_csv_path,
-            "--pdb-file",
-            protein_path,
-            "--struct-dir",
-            struct_dir,
-            "--output-dir",
-            tmp_path,
-            "--pymol-save",
-            tmp_path / "file.pse",
-            "--chain",
-            "both",
-            "--color-by-rmsd",
-        ],
-    )
-    assert click_success(result)
-
-
-@pytest.mark.skipif(os.getenv("SKIP_EXPENSIVE_TESTS"), reason="Expensive tests skipped")
-def test_struct_alignment_cfold_dir(blast_csv_path, protein_path, cfold_dir, tmp_path):
-    runner = CliRunner()
-    cfold_dir, _ = cfold_dir
-    result = runner.invoke(
-        cli,
-        [
-            "struct-alignment",
-            "-f",
-            blast_csv_path,
-            "--pdb-file",
-            protein_path,
-            "--cfold-results",
-            cfold_dir,
-            "--output-dir",
-            tmp_path,
-            "--pymol-save",
-            tmp_path / "file.pse",
-            "--chain",
-            "both",
-            "--color-by-rmsd",
-            "--cf-format",
-            "alphafold2_multimer_v3",
-        ],
-    )
-    assert click_success(result)
-
-
-@pytest.mark.skipif(os.getenv("SKIP_EXPENSIVE_TESTS"), reason="Expensive tests skipped")
-def test_fitness_alignment_pairwise(blast_csv_path, protein_path, tmp_path):
-    runner = CliRunner()
-    result = runner.invoke(
-        cli,
-        [
-            "fitness-alignment",
-            "-t",
-            "pwise",
-            "--pdb-file",
-            protein_path,
-            "--pdb-align",
-            protein_path,
-            "--pdb-label",
-            "ref,pdb",
-            "--pymol-save",
-            tmp_path / "file.pse",
-        ],
-    )
-    assert click_success(result)
-
-
-@pytest.mark.skipif(os.getenv("SKIP_EXPENSIVE_TESTS"), reason="Expensive tests skipped")
-def test_fitness_alignment_fasta(
-    blast_csv_path, fasta_alignment_path, protein_path, protein_mers_path, tmp_path
-):
-    runner = CliRunner()
-    result = runner.invoke(
-        cli,
-        [
-            "fitness-alignment",
-            "-t",
-            "fasta",
-            "--pdb-file",
-            protein_path,
-            "--pdb-align",
-            protein_mers_path,
-            "--pdb-label",
-            "ref,pdb",
-            "--pymol-save",
-            tmp_path / "file.pse",
-            "--fasta-a",
-            fasta_alignment_path,
-            "--fasta-b",
-            fasta_alignment_path,
-            "--fasta-sel",
-            "0,4",
-        ],
-    )
-    assert click_success(result)
